hiver_graphql::engine

Trait OutputType

pub trait OutputType: Send + Sync {
    // Required methods
    fn type_name() -> Cow<'static, str>;
    fn create_type_info(registry: &mut Registry) -> String;
    fn resolve(
        &self,
        ctx: &ContextBase<'_, &Positioned<SelectionSet>>,
        field: &Positioned<Field>,
    ) -> impl Future<Output = Result<ConstValue, ServerError>> + Send;

    // Provided methods
    fn qualified_type_name() -> String { ... }
    fn introspection_type_name(&self) -> Cow<'static, str> { ... }
}

Re-exported from async-graphql for convenience. 从 async-graphql 重新导出，方便使用。 Represents a GraphQL output type.

Required Methods

fn type_name() -> Cow<'static, str>

Type the name.

fn create_type_info(registry: &mut Registry) -> String

Create type information in the registry and return qualified typename.

fn resolve( &self, ctx: &ContextBase<'_, &Positioned<SelectionSet>>, field: &Positioned<Field>, ) -> impl Future<Output = Result<ConstValue, ServerError>> + Send

Resolve an output value to async_graphql::Value.

Provided Methods

fn qualified_type_name() -> String

Qualified typename.

fn introspection_type_name(&self) -> Cow<'static, str>

Introspection type name

Is the return value of field __typename, the interface and union should return the current type, and the others return Type::type_name.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.